GUACAMOLE-422: Remove duplicate code and migrate handshake to user handlers.
