+++ src/libguac/guacamole/user.h
@@ -506,79 +506,6 @@ struct guac_user {
 };
 
 /**
- * Handler for Guacamole protocol opcode specific to the handshake that
- * happens between client and server at the beginning of the connection.  The
- * handler will be invoked when the matching opcode is received during the
- * handshake process.
- * 
- * @param user
- *     The user that initiated the handshake.
- * 
- * @param parser
- *     The parser allocated for parsing the data provided by the client.
- * 
- * @param timeout
- *     The timeout, in microseconds, for parsing the value.
- * 
- * @return
- *     Zero if the handshake instruction is successfully parsed; otherwise
- *     false.
- */
-typedef int __guac_handshake_handler(guac_user* user, int argc, char** argv);
-
-/**
- * Structure that maps opcodes received during the handshake phase of the
- * connection to callback functions used when those opcodes are received.
- */
-typedef struct __guac_handshake_mapping {
-    
-    /**
-     * The instruction opcode which maps to the handler.
-     */
-    char* opcode;
-    
-    /**
-     * The handler function used when specified opcode is received.
-     */
-    __guac_handshake_handler* handler;
-    
-} __guac_handshake_mapping;
-
-/**
- * Internal handler function that is called when the size instruction is
- * received during the handshake process.
- */
-__guac_handshake_handler __guac_handshake_size_handler;
-
-/**
- * Internal handler function that is called when the audio instruction is
- * received during the handshake process, specifying the audio mimetypes
- * available to the client.
- */
-__guac_handshake_handler __guac_handshake_audio_handler;
-
-/**
- * Internal handler function that is called when the video instruction is
- * received during the handshake process, specifying the video mimetypes
- * available to the client.
- */
-__guac_handshake_handler __guac_handshake_video_handler;
-
-/**
- * Internal handler function that is called when the image instruction is
- * received during the handshake process, specifying the image mimetypes
- * available to the client.
- */
-__guac_handshake_handler __guac_handshake_image_handler;
-
-/**
- * Internal handler function that is called when the timezone instruction is
- * received during the handshake process, specifying the timezone of the
- * client.
- */
-__guac_handshake_handler __guac_handshake_timezone_handler;
-
-/**
  * Allocates a new, blank user, not associated with any specific client or
  * socket.
  *
@@ -616,32 +543,6 @@ void guac_user_free(guac_user* user);
  *     connection from being handled properly.
  */
 int guac_user_handle_connection(guac_user* user, int usec_timeout);
-
-/**
- * Call the appropriate handler defined by the given user for the given
- * instruction. A comparison is made between the instruction opcode and the
- * initial handler lookup table defined in user-handlers.c. The intial handlers
- * will in turn call the user's handler (if defined).
- *
- * @param user
- *     The user whose handlers should be called.
- *
- * @param opcode
- *     The opcode of the instruction to pass to the user via the appropriate
- *     handler.
- *
- * @param argc
- *     The number of arguments which are part of the instruction.
- *
- * @param argv
- *     An array of all arguments which are part of the instruction.
- *
- * @return
- *     Non-negative if the instruction was handled successfully, or negative
- *     if an error occurred.
- */
-int guac_user_handle_instruction(guac_user* user, const char* opcode,
-        int argc, char** argv);
 
 /**
  * Allocates a new stream. An arbitrary index is automatically assigned
