+++ src/libguac/user.c
@@ -167,24 +167,6 @@ void guac_user_free_object(guac_user* user, guac_object* object) {
 
 }
 
-int guac_user_handle_instruction(guac_user* user, const char* opcode, int argc, char** argv) {
-
-    /* For each defined instruction */
-    __guac_instruction_handler_mapping* current = __guac_instruction_handler_map;
-    while (current->opcode != NULL) {
-
-        /* If recognized, call handler */
-        if (strcmp(opcode, current->opcode) == 0)
-            return current->handler(user, argc, argv);
-
-        current++;
-    }
-
-    /* If unrecognized, ignore */
-    return 0;
-
-}
-
 void guac_user_stop(guac_user* user) {
     user->active = 0;
 }
